+++ apiserver/httpserver/auth.go
@@ -23,7 +23,7 @@ import (
 	"github.com/emicklei/go-restful/v3"
 	"github.com/golang/protobuf/proto"
 
-	"github.com/polarismesh/polaris-server/apiserver/httpserver/handler"
+	httpcommon "github.com/polarismesh/polaris-server/apiserver/httpserver/http"
 	api "github.com/polarismesh/polaris-server/common/api/v1"
 	"github.com/polarismesh/polaris-server/common/utils"
 )
@@ -66,7 +66,7 @@ func (h *HTTPServer) GetAuthServer(ws *restful.WebService) error {
 
 // AuthStatus auth status
 func (h *HTTPServer) AuthStatus(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	checker := h.authServer.GetAuthChecker()
 
@@ -83,7 +83,7 @@ func (h *HTTPServer) AuthStatus(req *restful.Request, rsp *restful.Response) {
 
 // Login 登陆函数
 func (h *HTTPServer) Login(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	loginReq := &api.LoginRequest{}
 
@@ -98,7 +98,7 @@ func (h *HTTPServer) Login(req *restful.Request, rsp *restful.Response) {
 
 // CreateUsers 批量创建用户
 func (h *HTTPServer) CreateUsers(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var users UserArr
 
@@ -117,7 +117,7 @@ func (h *HTTPServer) CreateUsers(req *restful.Request, rsp *restful.Response) {
 
 // UpdateUser 更新用户
 func (h *HTTPServer) UpdateUser(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	user := &api.User{}
 
@@ -132,7 +132,7 @@ func (h *HTTPServer) UpdateUser(req *restful.Request, rsp *restful.Response) {
 
 // UpdateUserPassword 更新用户
 func (h *HTTPServer) UpdateUserPassword(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	user := &api.ModifyUserPassword{}
 
@@ -147,7 +147,7 @@ func (h *HTTPServer) UpdateUserPassword(req *restful.Request, rsp *restful.Respo
 
 // DeleteUsers 批量删除用户
 func (h *HTTPServer) DeleteUsers(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var users UserArr
 
@@ -166,9 +166,9 @@ func (h *HTTPServer) DeleteUsers(req *restful.Request, rsp *restful.Response) {
 
 // GetUsers 查询用户
 func (h *HTTPServer) GetUsers(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
 
 	handler.WriteHeaderAndProto(h.authServer.GetUsers(ctx, queryParams))
@@ -176,8 +176,8 @@ func (h *HTTPServer) GetUsers(req *restful.Request, rsp *restful.Response) {
 
 // GetUserToken 获取这个用户所关联的所有用户组列表信息，支持翻页
 func (h *HTTPServer) GetUserToken(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
-	queryParams := utils.ParseQueryParams(req)
+	handler := &httpcommon.Handler{req, rsp}
+	queryParams := httpcommon.ParseQueryParams(req)
 
 	user := &api.User{
 		Id: utils.NewStringValue(queryParams["id"]),
@@ -188,7 +188,7 @@ func (h *HTTPServer) GetUserToken(req *restful.Request, rsp *restful.Response) {
 
 // UpdateUserToken 更改用户的token
 func (h *HTTPServer) UpdateUserToken(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	user := &api.User{}
 
@@ -203,7 +203,7 @@ func (h *HTTPServer) UpdateUserToken(req *restful.Request, rsp *restful.Response
 
 // ResetUserToken 重置用户 token
 func (h *HTTPServer) ResetUserToken(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	user := &api.User{}
 
@@ -218,7 +218,7 @@ func (h *HTTPServer) ResetUserToken(req *restful.Request, rsp *restful.Response)
 
 // CreateGroup 创建用户组
 func (h *HTTPServer) CreateGroup(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	group := &api.UserGroup{}
 
@@ -233,7 +233,7 @@ func (h *HTTPServer) CreateGroup(req *restful.Request, rsp *restful.Response) {
 
 // UpdateGroups 更新用户组
 func (h *HTTPServer) UpdateGroups(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var groups ModifyGroupArr
 
@@ -252,7 +252,7 @@ func (h *HTTPServer) UpdateGroups(req *restful.Request, rsp *restful.Response) {
 
 // DeleteGroups 删除用户组
 func (h *HTTPServer) DeleteGroups(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var groups GroupArr
 
@@ -271,9 +271,9 @@ func (h *HTTPServer) DeleteGroups(req *restful.Request, rsp *restful.Response) {
 
 // GetGroups 获取用户组列表
 func (h *HTTPServer) GetGroups(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
 
 	handler.WriteHeaderAndProto(h.authServer.GetGroups(ctx, queryParams))
@@ -281,9 +281,9 @@ func (h *HTTPServer) GetGroups(req *restful.Request, rsp *restful.Response) {
 
 // GetGroup 获取用户组详细
 func (h *HTTPServer) GetGroup(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
 
 	group := &api.UserGroup{
@@ -295,9 +295,9 @@ func (h *HTTPServer) GetGroup(req *restful.Request, rsp *restful.Response) {
 
 // GetGroupToken 获取用户组 token
 func (h *HTTPServer) GetGroupToken(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
 
 	group := &api.UserGroup{
@@ -309,7 +309,7 @@ func (h *HTTPServer) GetGroupToken(req *restful.Request, rsp *restful.Response) 
 
 // UpdateGroupToken 更新用户组 token
 func (h *HTTPServer) UpdateGroupToken(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	group := &api.UserGroup{}
 
@@ -324,7 +324,7 @@ func (h *HTTPServer) UpdateGroupToken(req *restful.Request, rsp *restful.Respons
 
 // ResetGroupToken 重置用户组 token
 func (h *HTTPServer) ResetGroupToken(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	group := &api.UserGroup{}
 
@@ -339,7 +339,7 @@ func (h *HTTPServer) ResetGroupToken(req *restful.Request, rsp *restful.Response
 
 // CreateStrategy 创建鉴权策略
 func (h *HTTPServer) CreateStrategy(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	strategy := &api.AuthStrategy{}
 
@@ -354,7 +354,7 @@ func (h *HTTPServer) CreateStrategy(req *restful.Request, rsp *restful.Response)
 
 // UpdateStrategies 更新鉴权策略
 func (h *HTTPServer) UpdateStrategies(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var strategies ModifyStrategyArr
 
@@ -373,7 +373,7 @@ func (h *HTTPServer) UpdateStrategies(req *restful.Request, rsp *restful.Respons
 
 // DeleteStrategies 批量删除鉴权策略
 func (h *HTTPServer) DeleteStrategies(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var strategies StrategyArr
 
@@ -392,9 +392,9 @@ func (h *HTTPServer) DeleteStrategies(req *restful.Request, rsp *restful.Respons
 
 // GetStrategies 批量获取鉴权策略
 func (h *HTTPServer) GetStrategies(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
 
 	handler.WriteHeaderAndProto(h.authServer.GetStrategies(ctx, queryParams))
@@ -402,9 +402,9 @@ func (h *HTTPServer) GetStrategies(req *restful.Request, rsp *restful.Response) 
 
 // GetStrategy 获取鉴权策略详细
 func (h *HTTPServer) GetStrategy(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
 
 	strategy := &api.AuthStrategy{
@@ -416,9 +416,9 @@ func (h *HTTPServer) GetStrategy(req *restful.Request, rsp *restful.Response) {
 
 // GetPrincipalResources 获取鉴权策略详细
 func (h *HTTPServer) GetPrincipalResources(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
 
 	handler.WriteHeaderAndProto(h.authServer.GetPrincipalResources(ctx, queryParams))
