+++ apiserver/httpserver/http/handler.go
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package utils
+package http
 
 import (
 	"context"
@@ -27,24 +27,15 @@ import (
 	"github.com/emicklei/go-restful/v3"
 	"github.com/golang/protobuf/jsonpb"
 	"github.com/golang/protobuf/proto"
+	"github.com/golang/protobuf/ptypes/wrappers"
 	"go.uber.org/zap"
 
+	"github.com/polarismesh/polaris-server/apiserver/httpserver/i18n"
 	api "github.com/polarismesh/polaris-server/common/api/v1"
 	apiv2 "github.com/polarismesh/polaris-server/common/api/v2"
 	"github.com/polarismesh/polaris-server/common/log"
+	"github.com/polarismesh/polaris-server/common/utils"
 )
-
-// parseQueryParams 解析并获取HTTP的query params
-func ParseQueryParams(req *restful.Request) map[string]string {
-	queryParams := make(map[string]string)
-	for key, value := range req.Request.URL.Query() {
-		if len(value) > 0 {
-			queryParams[key] = value[0] // 暂时默认只支持一个查询
-		}
-	}
-
-	return queryParams
-}
 
 // Handler HTTP请求/回复处理器
 type Handler struct {
@@ -78,16 +69,16 @@ func (h *Handler) postParseMessage(requestID string) (context.Context, error) {
 	platformID := h.Request.HeaderParameter("Platform-Id")
 	platformToken := h.Request.HeaderParameter("Platform-Token")
 	token := h.Request.HeaderParameter("Polaris-Token")
-	authToken := h.Request.HeaderParameter(HeaderAuthTokenKey)
+	authToken := h.Request.HeaderParameter(utils.HeaderAuthTokenKey)
 	ctx := context.Background()
-	ctx = context.WithValue(ctx, StringContext("request-id"), requestID)
-	ctx = context.WithValue(ctx, StringContext("platform-id"), platformID)
-	ctx = context.WithValue(ctx, StringContext("platform-token"), platformToken)
+	ctx = context.WithValue(ctx, utils.StringContext("request-id"), requestID)
+	ctx = context.WithValue(ctx, utils.StringContext("platform-id"), platformID)
+	ctx = context.WithValue(ctx, utils.StringContext("platform-token"), platformToken)
 	if token != "" {
-		ctx = context.WithValue(ctx, StringContext("polaris-token"), token)
+		ctx = context.WithValue(ctx, utils.StringContext("polaris-token"), token)
 	}
 	if authToken != "" {
-		ctx = context.WithValue(ctx, ContextAuthTokenKey, authToken)
+		ctx = context.WithValue(ctx, utils.ContextAuthTokenKey, authToken)
 	}
 
 	var operator string
@@ -101,7 +92,7 @@ func (h *Handler) postParseMessage(requestID string) (context.Context, error) {
 	if staffName := h.Request.HeaderParameter("Staffname"); staffName != "" {
 		operator = staffName
 	}
-	ctx = context.WithValue(ctx, StringContext("operator"), operator)
+	ctx = context.WithValue(ctx, utils.StringContext("operator"), operator)
 
 	return ctx, nil
 }
@@ -122,17 +113,18 @@ func (h *Handler) ParseHeaderContext() context.Context {
 	platformID := h.Request.HeaderParameter("Platform-Id")
 	platformToken := h.Request.HeaderParameter("Platform-Token")
 	token := h.Request.HeaderParameter("Polaris-Token")
-	authToken := h.Request.HeaderParameter(HeaderAuthTokenKey)
+	authToken := h.Request.HeaderParameter(utils.HeaderAuthTokenKey)
 
 	ctx := context.Background()
-	ctx = context.WithValue(ctx, StringContext("request-id"), requestID)
-	ctx = context.WithValue(ctx, StringContext("platform-id"), platformID)
-	ctx = context.WithValue(ctx, StringContext("platform-token"), platformToken)
+	ctx = context.WithValue(ctx, utils.StringContext("request-id"), requestID)
+	ctx = context.WithValue(ctx, utils.StringContext("platform-id"), platformID)
+	ctx = context.WithValue(ctx, utils.StringContext("platform-token"), platformToken)
+	ctx = context.WithValue(ctx, utils.ContextClientAddress, h.Request.Request.RemoteAddr)
 	if token != "" {
-		ctx = context.WithValue(ctx, StringContext("polaris-token"), token)
+		ctx = context.WithValue(ctx, utils.StringContext("polaris-token"), token)
 	}
 	if authToken != "" {
-		ctx = context.WithValue(ctx, ContextAuthTokenKey, authToken)
+		ctx = context.WithValue(ctx, utils.ContextAuthTokenKey, authToken)
 	}
 
 	var operator string
@@ -146,20 +138,20 @@ func (h *Handler) ParseHeaderContext() context.Context {
 	if staffName := h.Request.HeaderParameter("Staffname"); staffName != "" {
 		operator = staffName
 	}
-	ctx = context.WithValue(ctx, StringContext("operator"), operator)
+	ctx = context.WithValue(ctx, utils.StringContext("operator"), operator)
 
 	return ctx
 }
 
 // WriteHeader 仅返回Code
 func (h *Handler) WriteHeader(polarisCode uint32, httpStatus int) {
-	requestID := h.Request.HeaderParameter(PolarisRequestID)
-	h.Request.SetAttribute(PolarisCode, polarisCode) // api统计的时候，用该code
+	requestID := h.Request.HeaderParameter(utils.PolarisRequestID)
+	h.Request.SetAttribute(utils.PolarisCode, polarisCode) // api统计的时候，用该code
 
 	// 对于非200000的返回，补充实际的code到header中
 	if polarisCode != api.ExecuteSuccess {
-		h.Response.AddHeader(PolarisCode, fmt.Sprintf("%d", polarisCode))
-		h.Response.AddHeader(PolarisMessage, api.Code2Info(polarisCode))
+		h.Response.AddHeader(utils.PolarisCode, fmt.Sprintf("%d", polarisCode))
+		h.Response.AddHeader(utils.PolarisMessage, api.Code2Info(polarisCode))
 	}
 	h.Response.AddHeader("Request-Id", requestID)
 	h.Response.WriteHeader(httpStatus)
@@ -167,23 +159,22 @@ func (h *Handler) WriteHeader(polarisCode uint32, httpStatus int) {
 
 // WriteHeaderAndProto 返回Code和Proto
 func (h *Handler) WriteHeaderAndProto(obj api.ResponseMessage) {
-	requestID := h.Request.HeaderParameter(PolarisRequestID)
-	h.Request.SetAttribute(PolarisCode, obj.GetCode().GetValue())
+	requestID := h.Request.HeaderParameter(utils.PolarisRequestID)
+	h.Request.SetAttribute(utils.PolarisCode, obj.GetCode().GetValue())
 	status := api.CalcCode(obj)
 
 	if status != http.StatusOK {
 		log.Error(obj.String(), zap.String("request-id", requestID))
 	}
 	if code := obj.GetCode().GetValue(); code != api.ExecuteSuccess {
-		h.Response.AddHeader(PolarisCode, fmt.Sprintf("%d", code))
-		h.Response.AddHeader(PolarisMessage, api.Code2Info(code))
+		h.Response.AddHeader(utils.PolarisCode, fmt.Sprintf("%d", code))
+		h.Response.AddHeader(utils.PolarisMessage, api.Code2Info(code))
 	}
-
-	h.Response.AddHeader(PolarisRequestID, requestID)
+	h.Response.AddHeader(utils.PolarisRequestID, requestID)
 	h.Response.WriteHeader(status)
 
 	m := jsonpb.Marshaler{Indent: " ", EmitDefaults: true}
-	err := m.Marshal(h.Response, obj)
+	err := m.Marshal(h.Response, h.i18n(obj))
 	if err != nil {
 		log.Error(err.Error(), zap.String("request-id", requestID))
 	}
@@ -191,19 +182,19 @@ func (h *Handler) WriteHeaderAndProto(obj api.ResponseMessage) {
 
 // WriteHeaderAndProto 返回Code和Proto
 func (h *Handler) WriteHeaderAndProtoV2(obj apiv2.ResponseMessage) {
-	requestID := h.Request.HeaderParameter(PolarisRequestID)
-	h.Request.SetAttribute(PolarisCode, obj.GetCode())
+	requestID := h.Request.HeaderParameter(utils.PolarisRequestID)
+	h.Request.SetAttribute(utils.PolarisCode, obj.GetCode())
 	status := apiv2.CalcCode(obj)
 
 	if status != http.StatusOK {
 		log.Error(obj.String(), zap.String("request-id", requestID))
 	}
 	if code := obj.GetCode(); code != api.ExecuteSuccess {
-		h.Response.AddHeader(PolarisCode, fmt.Sprintf("%d", code))
-		h.Response.AddHeader(PolarisMessage, api.Code2Info(code))
+		h.Response.AddHeader(utils.PolarisCode, fmt.Sprintf("%d", code))
+		h.Response.AddHeader(utils.PolarisMessage, api.Code2Info(code))
 	}
 
-	h.Response.AddHeader(PolarisRequestID, requestID)
+	h.Response.AddHeader(utils.PolarisRequestID, requestID)
 	h.Response.WriteHeader(status)
 
 	m := jsonpb.Marshaler{Indent: " ", EmitDefaults: true}
@@ -218,4 +209,35 @@ func HTTPResponse(req *restful.Request, rsp *restful.Response, code uint32) {
 	handler := &Handler{req, rsp}
 	resp := api.NewResponse(code)
 	handler.WriteHeaderAndProto(resp)
+}
+
+// i18n 依据resp.code进行国际化resp.info信息
+// 当与header中的信息不匹配时, 则使用原文, 后续通过新定义code的方式增量解决
+// 当header的msg 与 resp.info一致时, 根据resp.code国际化信息
+func (h *Handler) i18n(obj api.ResponseMessage) api.ResponseMessage {
+	hMsg := h.Response.Header().Get(utils.PolarisMessage)
+	info := obj.GetInfo()
+	if hMsg != info.GetValue() {
+		return obj
+	}
+	code := obj.GetCode()
+	msg, err := i18n.Translate(
+		code.GetValue(), h.Request.QueryParameter("lang"), h.Request.HeaderParameter("Accept-Language"))
+	if msg == "" || err != nil {
+		return obj
+	}
+	*info = wrappers.StringValue{Value: msg}
+	return obj
+}
+
+// parseQueryParams 解析并获取HTTP的query params
+func ParseQueryParams(req *restful.Request) map[string]string {
+	queryParams := make(map[string]string)
+	for key, value := range req.Request.URL.Query() {
+		if len(value) > 0 {
+			queryParams[key] = value[0] // 暂时默认只支持一个查询
+		}
+	}
+
+	return queryParams
 }
