+++ apiserver/httpserver/http/handler_test.go
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package handler
+package http
 
 import (
 	"net/http"
