+++ apiserver/httpserver/maintain_access.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/emicklei/go-restful/v3"
 
-	"github.com/polarismesh/polaris-server/apiserver/httpserver/handler"
+	httpcommon "github.com/polarismesh/polaris-server/apiserver/httpserver/http"
 	api "github.com/polarismesh/polaris-server/common/api/v1"
 	"github.com/polarismesh/polaris-server/common/utils"
 	"github.com/polarismesh/polaris-server/maintain"
@@ -53,7 +53,7 @@ func (h *HTTPServer) GetMaintainAccessServer() *restful.WebService {
 //           host，可选，查看指定host
 func (h *HTTPServer) GetServerConnections(req *restful.Request, rsp *restful.Response) {
 	ctx := initContext(req)
-	params := utils.ParseQueryParams(req)
+	params := httpcommon.ParseQueryParams(req)
 	connReq := maintain.ConnReq{
 		Protocol: params["protocol"],
 		Host:     params["host"],
@@ -70,7 +70,7 @@ func (h *HTTPServer) GetServerConnections(req *restful.Request, rsp *restful.Res
 // GetServerConnStats 获取连接缓存里面的统计信息
 func (h *HTTPServer) GetServerConnStats(req *restful.Request, rsp *restful.Response) {
 	ctx := initContext(req)
-	params := utils.ParseQueryParams(req)
+	params := httpcommon.ParseQueryParams(req)
 
 	var amount int = 0
 	if amountStr, ok := params["amount"]; ok {
@@ -133,7 +133,7 @@ func (h *HTTPServer) FreeOSMemory(req *restful.Request, rsp *restful.Response) {
 // CleanInstance 彻底清理flag=1的实例运维接口
 // 支持一个个清理
 func (h *HTTPServer) CleanInstance(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	instance := &api.Instance{}
 	ctx, err := handler.Parse(instance)
@@ -148,8 +148,8 @@ func (h *HTTPServer) CleanInstance(req *restful.Request, rsp *restful.Response) 
 // GetLastHeartbeat 获取实例，上一次心跳的时间
 func (h *HTTPServer) GetLastHeartbeat(req *restful.Request, rsp *restful.Response) {
 	ctx := initContext(req)
-	handler := &handler.Handler{req, rsp}
-	params := utils.ParseQueryParams(req)
+	handler := &httpcommon.Handler{req, rsp}
+	params := httpcommon.ParseQueryParams(req)
 	instance := &api.Instance{}
 	if id, ok := params["id"]; ok && id != "" {
 		instance.Id = utils.NewStringValue(id)
