+++ apiserver/httpserver/v2/naming_console_access.go
@@ -24,11 +24,10 @@ import (
 	"github.com/emicklei/go-restful/v3"
 	"github.com/golang/protobuf/proto"
 
-	"github.com/polarismesh/polaris-server/apiserver/httpserver/handler"
+	httpcommon "github.com/polarismesh/polaris-server/apiserver/httpserver/http"
 	apiv1 "github.com/polarismesh/polaris-server/common/api/v1"
 	apiv2 "github.com/polarismesh/polaris-server/common/api/v2"
 	"github.com/polarismesh/polaris-server/common/log"
-	"github.com/polarismesh/polaris-server/common/utils"
 )
 
 const (
@@ -88,7 +87,7 @@ func (h *HTTPServerV2) addDefaultAccess(ws *restful.WebService) {
 
 // CreateRoutings 创建规则路由
 func (h *HTTPServerV2) CreateRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var routings RoutingArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -107,7 +106,7 @@ func (h *HTTPServerV2) CreateRoutings(req *restful.Request, rsp *restful.Respons
 
 // DeleteRoutings 删除规则路由
 func (h *HTTPServerV2) DeleteRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var routings RoutingArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -131,7 +130,7 @@ func (h *HTTPServerV2) DeleteRoutings(req *restful.Request, rsp *restful.Respons
 
 // UpdateRoutings 修改规则路由
 func (h *HTTPServerV2) UpdateRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
 	var routings RoutingArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -155,9 +154,9 @@ func (h *HTTPServerV2) UpdateRoutings(req *restful.Request, rsp *restful.Respons
 
 // GetRoutings 查询规则路由
 func (h *HTTPServerV2) GetRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &handler.Handler{req, rsp}
+	handler := &httpcommon.Handler{req, rsp}
 
-	queryParams := utils.ParseQueryParams(req)
+	queryParams := httpcommon.ParseQueryParams(req)
 	ret := h.namingServer.GetRoutingConfigsV2(handler.ParseHeaderContext(), queryParams)
 	handler.WriteHeaderAndProtoV2(ret)
 }
